Got combat working. Simulation can only run for nearly 5 minutes before slowwing down unbearably.
Combat is pretty simple, if you're different factions in the same square, the one with the higer strength lives, others die. This is acomplished using 3 nested ArrayLists.

 Done on 2 hours of sleep and half a cup of coffee. Main logic worked after first compile, only had to add code to ignore your own faction memebers.

This needs to be sped up in order to be able to be run long term, or it needs to be made more turn based.(probablistic futures? Can an AI predict how this sytem evolves?)